requirements.txt
Flask>=2.0
mysql-connector-python>=8.0
python-dotenv>=0.20
Werkzeug>=2.0
